feat: add some new features; style: update .clang-format; test: add function and unit test lib and code

+++ build.sh
@@ -1,11 +1,18 @@
 ###
  # @Author: modnarshen
- # @Date: 2023.01.4 16:56:55
+ # @Date: 2023.01.04 16:56:55
  # @Note: Copyrights (c) 2022 modnarshen. All rights reserved.
 ### 
 
 build_dir="./.build/"
+target=${1-"use_cache"}
 
-rm -fr ${build_dir}
+if [[ "$target" == "all" ]]; then 
+    echo "BUILD with cache cleared"
+    rm -fr ${build_dir}
+else
+    echo "BUILD with cache"
+fi
+
 mkdir -p ${build_dir} && cd ${build_dir}
 cmake .. && make -j10
